Zero-out the stack slots of local variables in a method's prologue.
NoLeak-Life.cool exposed a problem where an uninitialized stack item
happens to point exactly to L2.

`.GenGC.check_copy` processes a pointer
if it points within [L2; Alloc.limit), and therefore picks up
the uninitialized stack item for processing.

As a legitimate object can only start at L2 + 8 to make space for
the 8 bytes of eye-catcher, `.GenGC.check_copy` does not find
an eye-catcher when inspecting the uninitialized stack item and
proceeds to abort the program.

We fix this by writing 0 into a method's local variables stack slots,
so there are no uninitialized items in the method's frame.