Copy a prototype object just before the constructor call.
NoLeak-InsertionSort.cool exposed a problem where we make a copy of a prototype object
to pass it as `this` to the constructor. But while evaluating all the other constructor
parameters the copy gets promoted to Old Area. As a result any assignments the constructor
makes end up recorded in the assignment stack. So that the referenced objects survive the
next minor collection whereas they shouldn't have otherwise.

The behavior described above is not really a GC bug or a bug at all. It just doesn't make
sense to make a prototype object copy so early.